Use global_position rather than position for moving cards around
Previously the code would take global_position in some places, and
position in other places, and mix them, with the effect that the
layout would completely break if the nodes weren't laid out at the top
level with the top-left corner at (0,0).

There are likely still issues with transforms that aren't just a
translation, but that's a more complicated task to ensure it's all
good. For now, just being able to use layouts is nice :)